[CleanFileTransfer POCKET beta 0.5b] Updated: 2020/05/22 THIS BUILD IS STILL A BETA. Performance may be slow or unstable. It shall not damage your computer tho.
About V0.5B:
- FINALLY it works as expected.
- Supports AGAIN Messaging while transfering files
- Maximum speed measured at 1.2 GB/s (yes, giga BYTE)
- Actual maximum speed based on tests: 2.5 MB/s (this is BETA, remember that folks!)
- Some memory leaks have been fixed!
- Now it does actually have a message if command is not found.
- Greater overall visual: You can actually see the bitrate! There are two modes: Actual bitrate (all data, when on standby); Transfer bitrate (when transfering files it will change to this mode to tell you exactly the speed of the transfer.
- Please don't test this with mobile data. I am not responsible of any huge data loss right now.
- Use this app as you wish!

+++ CleanFileTransfer POCKET/display_cmd.h
@@ -113,8 +113,8 @@ namespace Custom {
 	inline void DISPLAY<X, Y>::_setTitle(const std::string str)
 	{
 		for (size_t p = 0; p < X; p++) {
-			if (p < str.length()) setAt(p, 1, str[p]);
-			else setAt(p, 1, ' ');
+			if (p < str.length()) setAt(static_cast<int>(p), 1, str[p]);
+			else setAt(static_cast<int>(p), 1, ' ');
 		}
 	}
 
@@ -123,8 +123,8 @@ namespace Custom {
 	{
 		if (line_c < line_count && line_c >= 0) {
 			for (size_t p = 0; p < X; p++) {
-				if (p < str.length()) setAt(p, line_c + 3, str[p]);
-				else setAt(p, line_c + 3, ' ');
+				if (p < str.length()) setAt(static_cast<int>(p), line_c + 3, str[p]);
+				else setAt(static_cast<int>(p), line_c + 3, ' ');
 			}
 		}
 	}
@@ -166,7 +166,7 @@ namespace Custom {
 	{
 		m.lock();
 		for (int p = 2; p < static_cast<int>(X); p++) {
-			if (p - 2 < str.length()) setAt(p, Y - 5, str[p - 2]);
+			if (static_cast<size_t>(p) - 2 < str.length()) setAt(p, Y - 5, str[static_cast<size_t>(p) - 2]);
 			else setAt(p, Y - 5, ' ');
 		}
 		m.unlock();
@@ -226,7 +226,7 @@ namespace Custom {
 			auto& i = line_f[lc];
 			if (i) {
 				std::string ll = i();
-				_printAt(lc, ll);
+				_printAt(static_cast<int>(lc), ll);
 			}
 		}
 
@@ -260,8 +260,8 @@ namespace Custom {
 			}
 		}
 		for (size_t p = 0; p < X; p++) {
-			if (p < str.length()) setAt(p, Y - 1, str[p]);
-			else setAt(p, Y - 1, ' ');
+			if (p < str.length()) setAt(static_cast<int>(p), static_cast<int>(Y) - 1, str[p]);
+			else setAt(static_cast<int>(p), Y - 1, ' ');
 		}
 		m.unlock();
